read and save data routine

+++ public/q3/q3e3script.js
@@ -1,21 +1,32 @@
-// read from the localStorage saved as a string;
+// read from the localStorage saved as a string - to see if there are anything saved on the users coomputer
 let acctString = localStorage.getItem("accounts")
 if (!acctString) { accountList = {} } // initialize the variable to contain the list of accounts object
 else accountList = JSON.parse(acctString) // converts string into the correct data type in this case object
 
-const form = document.getElementById("dForm");
+const form = document.getElementById("dForm"); // get the HTML form from q3ge2Mendoza.html
 
-function showAccounts(event) {
-  //event.preventDefault();
-  //const form = event.target;
-  username = uname.value
-  accountList[username] = {};
-	accountList[username]["pword"] = pword.value;
-	accountList[username]["dob"] = dob.value;
-//	accountList["username"]["pd"] = pd[0].value;
-	accountList[username]["about"] = about.value;
-	
-	console.log(accountList)
-	acctString = JSON.stringify(accountList)
-	localStorage.setItem("accounts", acctString)
-}
+form.addEventListener("submit", function(e) { // assign an event handler of submit to the form
+    e.preventDefault(); // prevent page reload because forms gets submitted
+
+    if (confirm("Sure You Want To Save Your Work?")) {   
+        // use a predefined class to create an object of data
+        const data = new FormData(form);
+
+        // Convert to object
+        const obj = Object.fromEntries(data.entries()); // get all the data from the form
+        // place the object inside the accountList
+        // accountList is an object containing other objects with username as the key
+        accountList[obj.uname] = {};
+        for (let key in obj) { // go through the properties of the object and create another account
+            if (key != "uname") { 
+                accountList[obj.uname][key] = obj[key];
+            }
+        }
+        
+        console.log(accountList) // to check all the account information if it will be saved correctly
+        acctString = JSON.stringify(accountList) // convert object into string, as a requirement of localStorage
+        localStorage.setItem("accounts", acctString) // save on the user's computer
+    }
+  });
+
+
